+++ src/Futhark/CodeGen/Backends/MulticoreISPC.hs
@@ -71,7 +71,6 @@ compileProg version prog = do
       ( GC.compileProg'
           "ispc"
           version
-          mempty
           operations
           (ISPCState mempty mempty)
           ( do