+++ src/Futhark/IR/Prop/TypeOf.hs
@@ -119,6 +119,7 @@ basicOpType Assert {} =
   pure [Prim Unit]
 basicOpType (UpdateAcc _ v _ _) =
   pure <$> lookupType v
+basicOpType UserParam {} = pure [Prim int64]
 
 -- | The type of an expression.
 expExtType ::